move tests; correct equality test on formal args; optional comments in formal args
