+++ nix-parser/src/parser.rs
@@ -16,9 +16,6 @@ mod expr;
 mod partial;
 mod tokens;
 
-#[cfg(test)]
-mod tests;
-
 type IResult<'a, T> = nom::IResult<Tokens<'a>, T, Errors>;
 
 impl FromStr for Expr {
@@ -81,4 +78,43 @@ pub fn parse_source_file_partial(source: &str) -> Result<Partial<SourceFile>, Er
 
     partial.extend_errors(errors);
     Ok(partial)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    use crate::{
+        ast::{ExprFnDecl, FnDeclFormals, Formal},
+        nix,
+    };
+
+    #[test]
+    fn issue_6() {
+        let source = parse_source_file(
+            r#"
+{
+  baz = quux;
+  inherit ({foo}: foo) bar;
+
+}
+"#,
+        )
+        .unwrap();
+        eprintln!("{:?}", source);
+        let some_fn = ExprFnDecl::Formals(FnDeclFormals::new(
+            vec![Formal::new("foo".into(), None, Default::default())],
+            None,
+            None,
+            nix!(foo),
+            Default::default(),
+        ));
+        assert_eq!(
+            source.expr(),
+            &nix!({
+                baz = quux;
+                inherit (antiquote!(some_fn.into())) bar;
+            })
+        );
+    }
 }
